Added hibernate db calls instead of hardcoded data in payment-service

+++ Donesi.mi/user-service/src/main/java/com/donesi/mi/userservice/models/Payment.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 public class Payment {
     int id;
 
-    int amount;
+    Double amount;
 
     Currency currency;
 }
